requirements.txt
Flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.1.0
pytz==2023.3
Flask-Bootstrap==3.3.7.1